requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
SQLAlchemy==2.0.31
Alembic==1.13.2
bcrypt==4.1.3
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
WTForms==3.1.2
cryptography==42.0.8
pytest==8.2.2
pytest-flask==1.3.0
